Rename rerender-listener.ts -> keyboard-hack.ts
